MOve demandlib code to oemof/demand
