+++ src/oemof/demand/bdew/bdew_data/shlp_sigmoid_factors.csv
