+++ src/oemof/demand/bdew/heat_building.py
@@ -17,7 +17,7 @@ from math import ceil
 import numpy as np
 import pandas as pd
 
-from demandlib.tools import add_weekdays2df
+from oemof.demand.tools import add_weekdays2df
 
 
 class HeatBuilding:
