+++ src/oemof/demand/tools.py
