+++ src/oemof/demand/vdi/resources_weather/TRY2010_01_Jahr.dat
