+++ src/oemof/demand/vdi/resources_weather/TRY2010_06_Jahr.dat
