+++ src/oemof/demand/vdi/resources_weather/TRY2010_09_Jahr.dat
