+++ src/oemof/demand/vdi/resources_weather/TRY_polygons.geojson
